# 06_api_reference/openapi_spec.yaml
...
openapi: 3.1.0
info:
  title: Teivaka TFOS API
  description: |
    Teivaka Agricultural Transformation Operating System (Agri-TOS) API.

    **Platform:** Teivaka PTE LTD (Company No. 2025RC001894), Fiji
    **Founder:** Uraia Koroi Kama (Cody)

    ## Four Pillars
    - **Knowledge Base (KB):** Validated agronomic protocols — RAG-grounded, no hallucination
    - **Farm OS (TFOS):** Production cycles, harvests, labor, inputs, tasks, alerts
    - **AI Intelligence (TIS):** Teivaka Intelligence System — Voice + Text AI for farmers
    - **Community:** Shared protocols and benchmarks across Teivaka network

    ## Authentication
    - Bearer JWT (15-minute access token)
    - httpOnly refresh cookie (7-day refresh token)
    - Multi-tenancy: All operational endpoints enforce `tenant_id` via PostgreSQL RLS

    ## Primary Financial Metric
    **CoKG** (Cost per Kilogram) = (LaborCost + InputCost + OtherCost) / TotalHarvestQty_kg
    CoKG is always the **first field** in financial response objects.

    ## Fiji-Specific
    - Currency: FJD (Fijian Dollar), all monetary fields suffixed `_fjd`
    - Timezone: Pacific/Fiji (UTC+12)
    - WhatsApp-first communication via Twilio
    - Voice pipeline: Audio → Whisper → TIS → TFOS → Confirm (target: <5 seconds)

    ## TIS Rate Limits (Redis-enforced)
    | Tier | TIS Calls/Day |
    |------|---------------|
    | FREE | 5 |
    | BASIC | 20 |
    | PREMIUM | Unlimited |
    | CUSTOM | Unlimited |

    ## Real Farm Data (Reference)
...

...
servers:
...
security:
  - BearerAuth: []

# ============================================================
# COMPONENTS
# ============================================================
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT access token (15-minute expiry). Obtain from POST /auth/login.
        Format: Authorization: Bearer <token>
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          description: Response payload (varies by endpoint)
      required: [success, data]
...
    CycleCreate:
      type: object
      required: [production_id, planting_date]
      properties:
        production_id:
          type: string
          example: CRP-EGG
        planting_date:
          type: string
          format: date
          example: "2026-04-08"
        area_acres:
          type: number
          format: float
          nullable: true
          description: Defaults to PU area if not provided
        notes:
          type: string
          nullable: true
    CycleFinancials:
      type: object
      description: |
        Cycle P&L. cogk_fjd is ALWAYS the first field.
        CoKG = (total_labor_cost_fjd + total_input_cost_fjd + total_other_cost_fjd) / total_harvest_qty_kg
      properties:
        cogk_fjd:
          type: number
          format: float
          nullable: true
          description: "PRIMARY METRIC — Cost per Kilogram (FJD). Always first field."
          example: 1.86
        gross_margin_pct:
          type: number
          format: float
          nullable: true
          example: 34.2
        total_revenue_fjd:
          type: number
          format: float
          example: 840.00
        total_cost_fjd:
          type: number
          format: float
          example: 553.00
        net_profit_fjd:
          type: number
          format: float
          example: 287.00
        total_harvest_qty_kg:
          type: number
          format: float
          example: 294.0
        total_labor_cost_fjd:
          type: number
          format: float
          example: 480.00
        total_input_cost_fjd:
          type: number
          format: float
          example: 73.00
        total_other_cost_fjd:
          type: number
          format: float
          example: 0.00
        cycle_id:
          type: string
          example: CY-F001-26-002
        pu_id:
          type: string
          example: F001-PU002
        production_id:
          type: string
          example: CRP-EGG
        production_name:
          type: string
          example: Eggplant
        market_price_fjd:
          type: number
          format: float
          nullable: true
          example: 2.80
        margin_per_kg_fjd:
          type: number
          format: float
          nullable: true
          example: 0.94
        cogk_formula:
          type: string
          example: "FJD 553.00 / 294.0 kg = FJD 1.86/kg"
        period_start:
          type: string
          format: date
          example: "2026-02-21"
        period_end:
          type: string
          format: date
          nullable: true
        harvest_breakdown:
          type: array
          items:
            type: object
            properties:
              harvest_date:
                type: string
                format: date
              qty_kg:
                type: number
                format: float
              grade:
                type: string
              unit_price_fjd:
                type: number
                format: float
              revenue_fjd:
                type: number
                format: float
        computed_at:
          type: string
          format: date-time
          example: "2026-04-07T06:05:00+12:00"

    # ---- HARVESTS ----
    HarvestLog:
      type: object
      properties:
        harvest_id:
          type: string
          pattern: '^HRV-\d{8}-\d{3}$'
          example: HRV-20260407-001
        id:
          type: string
          format: uuid
          example: 550e8400-e29b-41d4-a716-446655440000
        pu_id:
          type: string
          example: F001-PU002
        cycle_id:
          type: string
          example: CY-F001-26-002
        qty_kg:
          type: number
          format: float
          example: 42.0
        grade:
          type: string
          enum: [A, B, C]
          example: A
        unit_price_fjd:
          type: number
          format: float
          example: 2.80
        total_value_fjd:
          type: number
          format: float
          example: 117.60
        harvest_date:
          type: string
          format: date
          example: "2026-04-07"
        compliance_status:
          type: string
          enum: [COMPLIANT, BLOCKED]
          example: COMPLIANT
        customer_id:
          type: string
          nullable: true
          example: CUS-001
        photo_url:
          type: string
          nullable: true
        cogk_updated:
          type: boolean
          example: true
        new_cogk_fjd:
          type: number
          format: float
          nullable: true
          example: 1.82
    HarvestCreate:
      type: object
      required: [harvest_date, qty_kg]
      properties:
        harvest_date:
          type: string
          format: date
          example: "2026-04-07"
        qty_kg:
          type: number
          format: float
          minimum: 0.1
          example: 42.0
        grade:
          type: string
          enum: [A, B, C]
          default: A
          example: A
        unit_price_fjd:
          type: number
          format: float
          nullable: true
          example: 2.80
        customer_id:
          type: string
          nullable: true
          example: CUS-001
        photo_url:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
...
paths:
...
  /production-units/{pu_id}/cycles:
    post:
      tags: [Cycles]
      summary: Create new production cycle
      description: |
        Creates a new cycle on a Production Unit.

        **Rotation validation runs automatically before creation.** If rotation is BLOCKED,
        returns HTTP 409 with details and suggested alternative crops.

        **Enforcement decisions:**
        - `APPROVED`: cycle created immediately
        - `BLOCKED`: HTTP 409 returned — FOUNDER can manually override
        - `OVERRIDE_REQUIRED`: FOUNDER can proceed, other roles receive 409

        A first scheduled task is automatically created upon cycle creation.
      security:
        - BearerAuth: []
      parameters:
        - name: pu_id
          in: path
          required: true
          schema:
            type: string
          example: F001-PU002
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CycleCreate'
            examples:
              approved_rotation:
                summary: Approved rotation — Long Bean after Eggplant
                value:
                  production_id: CRP-LBN
                  planting_date: "2026-04-08"
                  area_acres: 1.2
              blocked_rotation:
                summary: Blocked rotation — Eggplant after Eggplant (20 days only, need 60)
                value:
                  production_id: CRP-EGG
                  planting_date: "2026-04-08"
      responses:
        '201':
          description: Cycle created successfully
          content:
            application/json:
              example:
                success: true
                data:
                  cycle_id: CY-F001-26-003
                  pu_id: F001-PU002
                  production_id: CRP-LBN
                  production_name: Long Bean (Yardlong Bean)
                  planting_date: "2026-04-08"
                  stage: Establishment
                  validation_result:
                    allowed: true
                    enforcement_decision: APPROVED
                    rule_status: PREF
                  first_task:
                    task_id: TSK-20260408-001
                    title: Establishment check — CRP-LBN on F001-PU002
                    due_date: "2026-04-15"
                  stage_protocols:
                    - stage: Establishment
                      duration_days: 14
                      key_actions: [Weekly watering, Basal fertiliser application]
                    - stage: Vegetative
                      duration_days: 21
                      key_actions: [NPK every 14 days, Pest scouting]
                    - stage: Fruiting
                      duration_days: 30
                      key_actions: [Potassium top-up, Harvest monitoring]
        '409':
          description: Cycle blocked by rotation validation
          content:
            application/json:
              example:
                success: false
                error:
                  code: ROTATION_BLOCKED
                  message: "Cannot plant CRP-EGG on F001-PU002. Same-family rotation block: 60 days required, only 20 days have passed."
                  details:
                    rotation_blocked: true
                    rotation_validation:
                      allowed: false
                      enforcement_decision: BLOCKED
                      rule_status: BLOCK
                      min_rest_days: 60
                      days_since_last_harvest: 20
                      days_short: 40
                      rotation_key: CRP-EGG:CRP-EGG
                      earliest_allowed_date: "2026-05-17"
                    alternatives:
                      - production_id: CRP-FRB
                        production_name: French Beans
                        rule_status: PREF
                        min_rest_days: 0
                        ready_now: true
                      - production_id: CRP-LBN
                        production_name: Long Bean (Yardlong Bean)
                        rule_status: PREF
                        min_rest_days: 0
                        ready_now: true
                      - production_id: CRP-SCN
                        production_name: Sweet Corn
                        rule_status: OK
                        min_rest_days: 30
                        ready_now: false
                        ready_date: "2026-04-17"
  /cycles/{cycle_id}/financials:
    get:
      tags: [Cycles]
      summary: Cycle financial summary with CoKG
      description: |
        Returns P&L for a production cycle with **CoKG as the first field**.

        **CoKG formula:** (LaborCost + InputCost + OtherCost) / TotalHarvestQty_kg

        CoKG is null if no harvests have been recorded yet.
      security:
        - BearerAuth: []
      parameters:
        - name: cycle_id
          in: path
          required: true
          schema:
            type: string
          example: CY-F001-26-002
      responses:
        '200':
          description: Cycle financials with CoKG first
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/CycleFinancials'
              example:
                success: true
                data:
                  cogk_fjd: 1.86
                  gross_margin_pct: 34.2
                  total_revenue_fjd: 840.00
                  total_cost_fjd: 553.00
                  net_profit_fjd: 287.00
                  total_harvest_qty_kg: 294.0
                  total_labor_cost_fjd: 480.00
                  total_input_cost_fjd: 73.00
                  total_other_cost_fjd: 0.00
                  cycle_id: CY-F001-26-002
                  pu_id: F001-PU002
                  production_id: CRP-EGG
                  production_name: Eggplant
                  market_price_fjd: 2.80
                  margin_per_kg_fjd: 0.94
                  cogk_formula: "FJD 553.00 / 294.0 kg = FJD 1.86/kg"
                  period_start: "2026-02-21"
                  period_end: null
                  harvest_breakdown:
                    - harvest_date: "2026-03-15"
                      qty_kg: 82.0
                      grade: A
                      unit_price_fjd: 2.80
                      revenue_fjd: 229.60
                    - harvest_date: "2026-03-29"
                      qty_kg: 110.0
                      grade: A
                      unit_price_fjd: 2.80
                      revenue_fjd: 308.00
                    - harvest_date: "2026-04-05"
                      qty_kg: 102.0
                      grade: A
                      unit_price_fjd: 2.96
                      revenue_fjd: 301.92
                  computed_at: "2026-04-07T06:05:00+12:00"

  # ==================== ROTATION ====================
...
  /production-units/{pu_id}/harvests:
    post:
      tags: [Harvests]
      summary: Log harvest with chemical compliance check
      description: |
        Creates a harvest record. **Chemical compliance is checked automatically.**

        If any chemical applied to this PU has a withholding period that has not elapsed,
        the harvest is **BLOCKED (HTTP 409)** with the blocking chemicals listed.

        **Two enforcement layers:**
        1. API-level check here
        2. PostgreSQL trigger `trg_harvest_compliance_check` on `harvest_log` INSERT

        On success: CoKG is recalculated and returned as `new_cogk_fjd`.
      security:
        - BearerAuth: []
      parameters:
        - name: pu_id
          in: path
          required: true
          schema:
            type: string
          example: F001-PU002
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HarvestCreate'
            examples:
              compliant_harvest:
                summary: Compliant harvest — after withholding period
                value:
                  harvest_date: "2026-04-15"
                  qty_kg: 42.0
                  grade: A
                  unit_price_fjd: 2.80
                  customer_id: CUS-001
              blocked_harvest:
                summary: Blocked harvest — within Dimethoate 7-day WHD
                value:
                  harvest_date: "2026-04-07"
                  qty_kg: 42.0
                  grade: A
      responses:
        '201':
          description: Harvest logged successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/HarvestLog'
              example:
                success: true
                data:
                  harvest_id: HRV-20260415-001
                  id: 550e8400-e29b-41d4-a716-446655440001
                  pu_id: F001-PU002
                  cycle_id: CY-F001-26-002
                  qty_kg: 42.0
                  grade: A
                  unit_price_fjd: 2.80
                  total_value_fjd: 117.60
                  harvest_date: "2026-04-15"
                  compliance_status: COMPLIANT
                  customer_id: CUS-001
                  cogk_updated: true
                  new_cogk_fjd: 1.82
        '409':
          description: Harvest blocked — chemical compliance violation
          content:
            application/json:
              example:
                success: false
                error:
                  code: COMPLIANCE_VIOLATION
                  message: "Harvest blocked: Dimethoate 40% EC was applied 4 days ago. Withholding period: 7 days. Safe to harvest after 2026-04-10."
                  details:
                    compliance_blocked: true
                    blocking_chemicals:
                      - chemical_id: CHEM-001
                        chemical_name: Dimethoate 40% EC
                        application_date: "2026-04-03"
                        withholding_period_days: 7
                        days_since_application: 4
                        harvest_allowed_after: "2026-04-10"
                    earliest_safe_harvest_date: "2026-04-10"

    get:
      tags: [Harvests]
      summary: List harvests for a production unit
      security:
        - BearerAuth: []
      parameters:
        - name: pu_id
          in: path
          required: true
          schema:
            type: string
          example: F001-PU002
        - name: cycle_id
          in: query
          schema:
            type: string
        - name: from_date
          in: query
          schema:
            type: string
            format: date
        - name: to_date
          in: query
          schema:
            type: string
            format: date
        - name: page
          in: query
          schema:
            type: integer
            default: 1
      responses:
        '200':
          description: Harvest list
          content:
            application/json:
              example:
                success: true
                data:
                  - harvest_id: HRV-20260405-001
                    qty_kg: 102.0
                    grade: A
                    unit_price_fjd: 2.96
                    total_value_fjd: 301.92
                    harvest_date: "2026-04-05"
                    compliance_status: COMPLIANT
                    customer_id: CUS-001
                  - harvest_id: HRV-20260329-001
                    qty_kg: 110.0
                    grade: A
                    unit_price_fjd: 2.80
                    total_value_fjd: 308.00
                    harvest_date: "2026-03-29"
                    compliance_status: COMPLIANT
                meta:
                  total: 3
                  page: 1
                  page_size: 20
                  total_pages: 1

  # ==================== LABOR ====================
...